refactor: MessageCardSection.tsx의 상태관리 useMessageCard 훅으로 분리
