+++ src/pages/orderpage/MessageCardSection.tsx
@@ -1,9 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import styled from "@emotion/styled";
-import { useState, useEffect } from "react";
 import { Controller, useFormContext } from "react-hook-form";
+import { useMessageCard } from "@/pages/orderpage/hooks/useMessageCard";
 import { MESSAGE_CARD_LIST } from "@/mocks/messagecard_mock";
-import type { MessageCard } from "@/mocks/messagecard_mock";
 import type { FullOrderFormValues } from "@/utils/validator";
 
 interface Props {
@@ -11,21 +10,8 @@ interface Props {
 }
 
 const MessageCardSection = ({ error }: Props) => {
-  const { setValue, control } = useFormContext<FullOrderFormValues>();
-  const [selectedCard, setSelectedCard] = useState<MessageCard>(
-    MESSAGE_CARD_LIST[0]
-  );
-
-  useEffect(() => {
-    setValue("message", MESSAGE_CARD_LIST[0].defaultTextMessage);
-    setValue("messageCardId", String(MESSAGE_CARD_LIST[0].id));
-  }, [setValue]);
-
-  const handleSelect = (card: MessageCard) => {
-    setSelectedCard(card);
-    setValue("message", card.defaultTextMessage);
-    setValue("messageCardId", String(card.id));
-  };
+  const { control } = useFormContext<FullOrderFormValues>();
+  const { selectedCard, handleSelect } = useMessageCard();
 
   return (
     <Wrapper>
